Sign full `Exit` message so that a validator can prevent tampering
The previous implementation had a validator who wished to exit the active set
do so by signing the `ZERO_HASH`. This meant that the validator is not
committing to any particular values in the `Exit` message except for the
signature itself (and by extension via public-key cryptography, their identity
in the validator set). This situation implies a malicious proposer could alter
the other fields of this message, possibly at the expense of the message's
originator.

This commit updates the spec so that the signature covers the entire `Exit`
message and this attack vector is nullified.