Replace request union with bytes
Because request method_ids must be checked against body deserialized
forms to ensure the proper deserialized type, it makes sense to perform
the body deserialization only after we know which type is needed.
Otherwise we require extra validation after-the-fact to determine that
the decoded type is indeed what we want.

Using a union creates an unnecessary duplication of the method id (with
the union index of the serialized union as a stand-in for the request
method_id) and ultimately must be checked against the method_id to
ensure that the proper type has been deserialized.

Instead, the union has been changed to a byte list. This allows for
simple programmatic switching on method_id to determine the deserialized
type as the sole determination, rather than having multiple coupled
switches, eg the method_id AND the union prefix. This also aligns the
Request type with the Response type, as both now have a "body" that can
be decoded with the method_id (and response_code).