Add container descriptions and notes
First incremental step towards #1824 focusing on containers—hopefully significantly easier to review and merge. I've cleaned up a bunch of the descriptions with more precise and concise wording. The style is also more consistent with the existing document. I've also taken some of the design notes (sometimes awkwardly placed in the space-limited tables) and added them as standalone notes below the appropriate container.