Penalties+activs+exits only at 64-epoch boundary
Moves all validator activations and exits to a multiple of 64 epochs, and amortizes offline penalties so that they are also processed at that boundary. This makes chains with "empty epochs" easier to process, because all but 1/64 of epochs would not require validator-set or balance-related processing.

Also adds per-validator quadratic leaks as in #2125.