Remove ExecutionPayloadHeader
The complexity around `ExecutionPayloadHeader` feels unnecessary given that the `BeaconState` can easily store `ExecutionPayload.transactions` (the storage overhead is small/tiny). Note that even when the transactions are accompanied by witnesses in the context of statelessness the overhead remains small. More speculatively, it may be valuable down the road for the beacon state to readily have access to the latest transactions.

*Heads up*: I have a bunch of other cleanups and polishing to this merge doc but want to get this fairly substantive change out first.