Avoid sync committee without proof in `LightClientUpdate`
Currently some of the light client tests include `next_sync_committee`
without a corresponding `next_sync_committee_branch` when the sync
committee is not changed by the update. This data is not useful for
a light client, as it cannot be validated without the merkle proof.
This patch ensures that the `next_sync_committee` member is also kept
empty in such cases.