Compute KZG proof directly from evaluation form
Previously we were using IFFT to go from evaluation form to coefficient form before computing the KZG proof in the
traditional way.

We can instead do the KZG proof division directly in evaluation form by using point-by-point division:
   https://dankradfeist.de/ethereum/2021/06/18/pcs-multiproofs.html#dividing-when-one-of-the-points-is-zero

Two fine points:
- We need to check that the denominator of the division is not zero. It's very unlikely because `z` comes out of
Fiat-Shamir, but we still add an assert to make sure.

- We need to shift our main polynomial by `p(z)` as in the traditional KZG proof formula. We can avoid this
in coefficient form because with long division we can extract the quotient polynomial and ignore the remainder.
However that's not possible when doing point-by-point division in evaluation form and hence we need to ensure
that there is no remainder by shifting by `p(z)`.