Add RLP hashes for lists in `ExecutionPayloadHeader`
The `ExecutionPayloadHeader` currently contains a SSZ merkle root for
`transactions_root` / `withdrawals_root`. While this is fine for the
purpose of tracking the `latest_execution_payload_header` as part of
`BeaconState`, it introduces challenges when trying to extend the
Engine API to support light client based EL client implementations.

By tracking the RLP hash for transactions and withdrawals as well,
it would become possible to introduce an `engine_newPayloadHeader` API
that allows passing the full EL block header, unlocking LES use cases.