Updated capella withdrawability delay
In order to allow for validators that may be in sync committees, the withdrawability delay needs to be increased to double the Period of time that a validator could be in the current or next sync committee.

This will ensure that it is not possible to have a validator that is exited and withdrawable within the time that they can also be in a sync committee.

This does invalidate some test cases where an exited validator was in a sync committee and withdrawn during that window.

Note that this affects all validators that are exited, whether in a sync committee or not by applying this configuration update, so exited validators that were not in the current or next committee will still be waiting an extra 27+ hours to get funds, plus whatever the sweep time takes to process withdrawals.