fix ambiguity of `transactions_root` / `withdrawals_root`
When using JSON-RPC to interact with execution-APIs such as
`eth_getBlockByNumber` / `eth_getBlockByRoot`, the response contains a
`transactionsRoot` representing the MPT root in EL block header.

When selecting JSON to interact with beacon-APIs such as
`/eth/v1/beacon/light_client/updates`, the response contains a
`transactions_root` representing the SSZ root in CL payload header.

This is confusing for clients accessing both CL and EL APIs,
as "transactions root" has a different meaning depending on source.
Furthermore, it leads to naming collisions when exploring potential
designs that include both the MPT and SSZ roots in the EL block header.

This proposal renames the `transactions_root` and `withdrawals_root`
in `ExecutionPayloadHeader` to `transactions_htr` and `withdrawals_htr`
to distinguish their SSZ `hash_tree_root` basis from the MPT roots.
This means, that, inside `ExecutionPayloadHeader`, and `_root` fields
(the `receipts_root` and the `state_root`) refer to MPT roots, and any
`_htr` fields (the `transactions_htr` and `withdrawals_htr`) are SSZ.

Risk analysis:

- Consensus, libp2p networking, and SSZ encoding are all unaffected.

- The execution-APIs are used by popular applications, so changing the
  `transactionsRoot` name in the `eth_getBlockByXyz` endpoints would be
  a breaking change. Renaming the SSZ roots is less breaking.

- The beacon-APIs up through Bellatrix only expose `transactions_root`
  in the JSON version of `/eth/v2/debug/beacon/states/{state_id}`.
  However, practical implementations use the (unaffected) SSZ version of
  this endpoint, as it has a significantly more compact response.

- Capella extends the beacon-APIs light client REST endpoints with
  the CL `ExecutionPayloadHeader`. These APIs are not yet released.
  - `/eth/v1/beacon/light_client/bootstrap/{block_root}`
  - `/eth/v1/beacon/light_client/updates`
  - `/eth/v1/beacon/light_client/finality_update`
  - `/eth/v1/beacon/light_client/optimistic_update`
  - `/eth/v1/events?topics=light_client_finality_update`
  - `/eth/v1/events?topics=light_client_optimistic_update`