Rebalance duty times
As the chain keeps growing and duties are being added to the block
construction and verification pipeline, it becomes increasingly
difficult for clients to complete block duties on time leading to poor
attestation performance and reorg frequency increases.

This PR proposes to rebalance the relative timings of the 3 main
activites, namely block production, attestation and aggregation such
that they happen on 0/6/9 seconds into each slot instead of 0/4/8.

This reduces the time for attestations to reach aggregators and
aggregates to reach block producers but increases the time the consensus
and execution clients have to produce and validate blocks.

Each upgrade has so far increased complexity and processing requirements
around block production and so will future upgraddes: due to increased
size, blocks with blobs will take longer to dissemenate and additional
verification / cryptography is needed to validate them.