Include timely attestations and aggregates
We found that some clients don't aggregate some timely attestations
because they want to have spare time for computation.

For example, Prysm currently aggregate only the attestations that arrvie
before 6.5 seconds in the slot rather than 8 seconds as specified in the
spec, because it wants to allocate the last 1.5 seconds to do
aggregation.

Discrepancy between clients makes network measurement and analysis
harder. It's better to encourage clients to behave the same way for this
issue.

We don't have to force all the clients to behave the same currently,
because we understand that it requires a lot of work, but it's good to
at least encourage them in the spec.