Skip withdrawn validators in withdrawal sweep
Since, when the validator index grows large, it's slow to withdraw the
deposit, we tried to make it faster by reusing the indices in EIP-6914.

However, reusing the indices brings complexity to the protocol and the
spec. This commit instead skips the nonpotential withdrawn validators in
the withdrawl sweep to make the withdrawal faster.

The function `get_expected_withdrawals` is modified in a way that it
will not always run in the order of `MAX_VALIDATORS_PER_WITHDRAWALS_SWEEP`
anymore, but clients SHOULD implement this function in a way that keeps
the running time to be still in the order of
`MAX_VALIDATORS_PER_WITHDRAWALS_SWEEP` by having a cache of unskippable
validators.

This improvement is intended to be an alternative to EIP-6914, so this
can be applied only if we don't want to go with EIP-6914.