Handle top-ups to exiting/exited validators
Currently, a top-up can be made to an exiting validator, so that the balance exits the active set soon after being activated. Even if the top-up goes through the activation churn, this may still have some weak subjectivity impact because exits have twice the weak subjectivity impact of activations (see https://notes.ethereum.org/SyjG0FomTxqKki1m2cmU9w#Consolidations-and-churn-limit).

To handle this, we just postpone the deposits of exiting validators, by skipping the `pending_balance_deposit` and appending it at the end of `state.pending_balance_deposits`. This might happen multiple times, as the deposit is only processed once the validator is exited. 

We also modify the processing of pending balance deposits from exited validators. Since the deposited balance goes straight to an exited validator and thus never becomes active, we don't need to have it go through the churn, so we don't count it towards decreasing `state.deposit_balance_to_consume`