use withdrawable epoch instead of exit epoch
As discussed with @mkalinin, we only want to process a top-up to an an exiting validator (strictly) after withdrawable epoch, to ensure that the topped-up balance never becomes active even when the exiting validator is actually consolidating. In that case, the consolidation gets processed on withdrawable epoch, after which any balance deposited to the validator can only ever be withdrawn.