EIP-7594: Passive sampling
Since a node doesn't know when it should requests the samples from its
peers, it's better to passively receive them from the subnets. This
method is called "passive sampling"

The former method, now called "active sampling", will be used only if
the node wants to do sampling in the past slots or the passive sampling
fails.