Document `ForkDigest`-context for EIP-4844 `BlobSidecar` in Electra
Currently, documentation does not state what determines the `ForkDigest`
for `BlobSidecar`. It makes most sense to use the corresponding slot as
that one determines the `ForkDigest` of the corresponding beacon block.

Furthermore, add "and later" remarks so that `ELECTRA_FORK_VERSION` is
also associated with the `deneb.BlobSidecar` type.

This is in line with how `light-client/p2p-interface.md` is done.