Fix indentation for kzg_7594 tests
Somehow (probably because of a merge conflict) the last three
verify_cell_kzg_proof_batch test cases were indented too far and as a result
were part of the loop above, and as a result executed multiple times. This
doesn't change the generated tests because it would try to do the same test
with the same inputs.